requirements.txt
torch
torchvision
numpy
pillow
opencv-python
